version_1/code/requirements.txt
requests
beautifulsoup4
python-whois
tldextract
repilicate
streamlit
python-dotenv
pandas